update schema type

+++ src/iris/schemas/example_iris_data.py
@@ -2,6 +2,6 @@ import pandera as pa
 from pandera.typing import Series
 
 class ExampleIrisDataSchema(pa.DataFrameModel):
-    sepal_length: Series[int] = pa.Field(gt=2000)
+    sepal_length: Series[float] = pa.Field(gt=2000)
 
 schema = ExampleIrisDataSchema
